Basic Scaling and cleanup
