+++ src/ga/mod.rs
@@ -1,8 +1,10 @@
-// TODO: COPYRIGHT, USE & AUTHORS
-// TODO: RUST DOCS!
+// Copyright 2016 Revolution Solid & Contributors.
+// author(s): sysnett
+// rust-monster is licensed under a MIT License.
 pub mod ga_core;
 pub mod ga_population;
 pub mod ga_random;
 pub mod ga_scaling;
-pub mod ga_selectors;
 pub mod ga_simple;
+pub mod ga_selectors;
+pub mod ga_test;
